BAC WORKING

+++ speakEasy/Model/User.h
@@ -16,7 +16,7 @@
 
 /* Unique userId corresponding to our DB table */
 @property (strong) NSString *userID;
-@property (readwrite, assign) long score;
+@property (readwrite, assign) double score;
 @property (copy) NSString *name;
 @property (strong) NSMutableArray *friends;
 @property (strong) NSMutableArray *friendsScores;
